--- nc-news/src/components/About.jsx
@@ -1,1 +1,0 @@
-export default function About() {}
