+++ nc-news/src/components/Comments.jsx
@@ -1,9 +1,6 @@
 import { useState, useEffect } from "react";
 import { fetchArticleComments } from "../../utils/utils";
-import { Expandable } from "./styled-components/StyledComponents";
-import moment from "moment";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faThumbsUp, faThumbsDown } from "@fortawesome/free-solid-svg-icons";
+import CommentCard from "./CommentCard";
 
 export default function Comments({ currentArticle }) {
   const [comments, setComments] = useState([]);
@@ -16,29 +13,10 @@ export default function Comments({ currentArticle }) {
 
   return (
     <section>
-      <h3>Comments</h3>
-
-      <Expandable>
-        {comments.map((comment) => {
-          return (
-            <section key={comment.comment_id}>
-              <p>{comment.body}</p>
-              <br></br>
-              <p>
-                by {comment.author},{" "}
-                {moment(comment.created_at).format("MMM Do YYYY, h:mm a")}
-              </p>
-              <div>
-                <FontAwesomeIcon
-                  icon={+comment.votes < 0 ? faThumbsDown : faThumbsUp}
-                />{" "}
-                {comment.votes}
-              </div>
-              <br />
-            </section>
-          );
-        })}
-      </Expandable>
+      <header>
+        <h3>Comments</h3>
+      </header>
+      <CommentCard comments={comments} />
     </section>
   );
 }
