--- nc-news/src/components/Profile.jsx
@@ -1,1 +1,0 @@
-export default function Profile() {}
