+++ nc-news/src/components/SingleArticle.jsx
@@ -1,3 +1,4 @@
+import { useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCommentAlt, faThumbsUp } from "@fortawesome/free-solid-svg-icons";
 import moment from "moment";
