+++ nc-news/src/index.css
@@ -21,13 +21,13 @@
 .header {
   background-color: rgb(255, 253, 253);
   height: 5em;
-  padding: 0 2em;
+  padding: 0 1em;
   position: relative;
   z-index: 2;
 }
 
 .header__title h1 {
-  background: linear-gradient(to left, #a6bad9, #b45266);
+  background: linear-gradient(to left, #a6bad9, #e30909);
   -webkit-text-fill-color: transparent;
   -webkit-background-clip: text;
   font-size: 3em;
