+++ nc-news/src/index.jsx
@@ -1,8 +1,7 @@
 import NavBar from "./components/NavBar";
 import SideBar from "./components/SideBar";
 import AllArticles from "./components/AllArticles";
-import Profile from "./components/Profile";
-import About from "./components/About";
+
 import SingleArticle from "./components/SingleArticle";
 
-export { NavBar, SideBar, AllArticles, Profile, About, SingleArticle };
+export { NavBar, SideBar, AllArticles, SingleArticle };
